all: altera chamadas de getParameter() e getElement() para remover castings.

+++ mm/materials/classes/org/erp/custom/mm/materials/Common.java
@@ -14,6 +14,6 @@ public class Common {
     };
     
     public static final byte getMode(ViewData view) {
-        return (Byte)view.getParameter("mode");
+        return view.getParameter("mode");
     }
 }
